+++ Shared/MovappApp.swift
@@ -77,7 +77,7 @@ struct MovappApp: App {
         guard userDefaultsStore.disableSetIcon() == false else {
             return
         }
-        
+
         let app = UIApplication.shared
         let currentYear = Calendar.current.dateComponents([.year], from: Date()).year ?? 0